Started refactoring of rows. Moved NewsApp component to main folder.
